move filters to prefilter module
